fixed bugs

+++ consultingroom-service/src/main/java/com/feature4/consultingroom_service/dto/ConsultorioResponseDTO.java
@@ -1,7 +1,7 @@
 package com.feature4.consultingroom_service.dto;
 
 public record ConsultorioResponseDTO(
-        Integer id_consultorio,
+        Long id_consultorio,
         String numero_consultorio,
         String tipo,
         String estado,
